migrate test_ldapmod to pytest
